class structuring

+++ src/program/meta.hpp
@@ -10,7 +10,7 @@
 #define MODE_FIRSTHIT "firsthit"
 #define MODE_HELP "help"
 
-static const size_t mode_count = 5;
+static const unsigned int mode_count = 5;
 static const char *mode_names[] = {
    MODE_RENDER, MODE_SCENEINFO, MODE_PIXELRAY,
    MODE_FIRSTHIT, MODE_HELP
